Move Core public APIs into separate package
