+++ src/CLI/ConsoleEventLogger.cs
@@ -1,4 +1,4 @@
-﻿using Core;
+﻿using Core.API;
 using Core.Utils;
 
 namespace AMS2CM.CLI;
