+++ src/GUI/ModVM.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel;
 using Core;
+using Core.API;
 
 namespace AMS2CM.GUI;
