+++ tests/Core.Tests/ModInstallerIntegrationTest.cs
@@ -1,6 +1,7 @@
 using Core.Backup;
 using Core.Mods;
 using Core.State;
+using Core.Tests.Base;
 using Core.Utils;
 using FluentAssertions;
